+++ Source/CGFloatLiteral.swift
@@ -1,7 +1,4 @@
 //
-//  LiteralTypeExtension.swift
-//  SwiftyUtils
-//
 //  Created by Tom Baranes on 15/09/16.
 //  Copyright © 2016 Tom Baranes. All rights reserved.
 //
@@ -19,6 +16,7 @@ public extension IntegerLiteralType {
 }
 
 public extension FloatLiteralType {
+    
     // swiftlint:disable:next variable_name
     public var f: CGFloat {
         return CGFloat(self)